move managers to separate package
