+++ src/main/scala/nl.knaw.dans.easy.sword2/managers/ContainerManagerImpl.scala
@@ -13,11 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package nl.knaw.dans.easy.sword2
+package nl.knaw.dans.easy.sword2.managers
 
 import java.util
 
 import nl.knaw.dans.easy.sword2.DepositHandler._
+import nl.knaw.dans.easy.sword2.{ Authentication, DepositHandler, DepositId, DepositProperties, LdapAuthSettings, Settings, SwordConfig, SwordID }
 import nl.knaw.dans.lib.error._
 import nl.knaw.dans.lib.logging.DebugEnhancedLogging
 import org.swordapp.server._
