+++ src/main/scala/nl.knaw.dans.easy.sword2/managers/MediaResourceManagerImpl.scala
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package nl.knaw.dans.easy.sword2
+package nl.knaw.dans.easy.sword2.managers
 
 import java.util
