+++ src/main/scala/nl.knaw.dans.easy.sword2/managers/ServiceDocumentManagerImpl.scala
@@ -13,8 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package nl.knaw.dans.easy.sword2
+package nl.knaw.dans.easy.sword2.managers
 
+import nl.knaw.dans.easy.sword2.{ Authentication, Settings, SwordConfig }
 import nl.knaw.dans.lib.logging.DebugEnhancedLogging
 import org.swordapp.server._
